move optimism repo setup to own file
